+++ includes/globals.php
@@ -5,7 +5,7 @@
  */
 
 // Current CMS version
-const VERSION = 'Version 1.4.8 (&alpha;)';
+const VERSION = 'Version 1.4.9 (&alpha;)';
 
 /**
  * Display copyright on the admin dashboard.